tls: add stateless TLS10 kernel descriptors
These descriptors are part of Freescale's SDK caam driver
and are valid on platforms with 36/40-bit address pointers.

Supported cipher suites: aes-cbc & sha1.

For the moment ALWAYS sharing is used at decapsulation, due to
incorrect handling of AES's Decrypt form of the key.
SERIAL sharing support will be added in a future patch.

Change-Id: I76f091543c96f9da8b53a9dac895ff120a2cd6ee
Signed-off-by: Tudor Ambarus <tudor.ambarus@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/29111
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>