fix PDCP SNOW f8 + AES CMAC decapsulation
Due to the fact that SEC HW cannot process both SNOW f8 and
AES-CMAC algorithms at the same time (both are Class 1
algorithms), it will write the decrypted frame in the output
buffer and switch the input & the output buffers. This poses
an issue if the output buffer length is not equal to input
buffer length - 4B (ICV): SEC will read the full buffer length,
and perform AES-CMAC ICV check. This leads to ICV check failures.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>